Phase 3: Extract Import tab UI class
Refactoring changes:
- Created ui/ package for tab-specific UI components
- Extracted ImportTab class to ui/import_tab.py (217 lines)
- Removed import tab creation and methods from main file
- ImportTab is now a self-contained QWidget with all import functionality
- Handles cross-tab dependencies (clear_db_btn) via dependency injection

Benefits:
- Improved separation of concerns (UI logic separated by tab)
- Reduced main file from 2,815 → 2,570 lines (245 lines removed)
- Import functionality is now modular and testable
- Establishes pattern for extracting remaining tabs

All functionality preserved and syntax verified.

This completes Phase 3 with Import tab extraction. Future phases can
continue extracting remaining tabs (View, Sessions, Analytics,
Maintenance, Settings) using the same pattern.
